Move getDefaultFormat to LifecyclePlugin
