+++ packages/roosterjs-editor-types/lib/interface/EditorCore.ts
@@ -3,7 +3,6 @@ import EditorPlugin from './EditorPlugin';
 import NodePosition from './NodePosition';
 import { ChangeSource } from '../enum/ChangeSource';
 import { ColorTransformDirection } from '../enum/ColorTransformDirection';
-import { DefaultFormat } from '..';
 import { DOMEventHandler } from '../type/domEventHandler';
 import { GetContentMode } from '../enum/GetContentMode';
 import { InsertOption } from './InsertOption';
@@ -40,14 +39,6 @@ export type AttachDomEvent = (
     core: EditorCore,
     eventMap: Record<string, DOMEventHandler>
 ) => () => void;
-
-/**
- * Get or recalculate default format of editor
- * @param core The EditorCore object
- * @param forceRecalculate If true, recalcuate default format then return. Otherwise return current value directly
- * This is used when default format is changed, e.g. when dark mode state is changed
- */
-export type GetDefaultFormat = (core: EditorCore, forceRecalculate: boolean) => DefaultFormat;
 
 /**
  * Create a DocumentFragment for paste from a ClipboardData
@@ -199,14 +190,6 @@ export interface CoreApiMap {
      * @param beforeDispatch Optional callback function to be invoked when the DOM event is triggered before trigger plugin event
      */
     attachDomEvent: AttachDomEvent;
-
-    /**
-     * Get or recalculate default format of editor
-     * @param core The EditorCore object
-     * @param forceRecalculate If true, recalcuate default format then return. Otherwise return current value directly
-     * This is used when default format is changed, e.g. when dark mode state is changed
-     */
-    getDefaultFormat: GetDefaultFormat;
 
     /**
      * Create a DocumentFragment for paste from a ClipboardData
